refactored runcmd functions
