+++ report.txt
@@ -1,0 +1,38 @@
+
+report:
+step: update submodules
+  command: 
+  duration: 0.00000119 seconds
+
+step: update submodules
+  command: 
+  duration: 0.00000072 seconds
+
+step: update submodules
+  command: 
+  duration: 0.00000048 seconds
+
+step: update submodules
+  command: 
+  duration: 0.00000048 seconds
+
+step: update submodules
+  command: git add .
+  duration: 0.00234699 seconds
+
+step: update submodules
+  command: 
+  duration: 0.00000143 seconds
+
+step: update submodules
+  command: git commit -m use --force-with-lease instead of --force
+  duration: 0.01812053 seconds
+  output: [dev 619a8d7] use --force-with-lease instead of --force
+ 2 files changed, 14 insertions(+), 11 deletions(-)
+
+
+step: update submodules
+  command: git push
+  duration: 4.77573442 seconds
+
+total duration: 4.90052438 seconds
